Two changes: WCS and fixing a weight bug in loading stars.
Changed wcs to affine wcs if images have origins. This helps with
keeping track of the wcs coordinates. I think. Unless I broke something
not in the tests.

Then, weights need to be 1, not 0, when loading stars and not their
images. If 0, then we won't use any of the pixels, so you can't actually
use the drawn stars. That's bad.